MAINT: move publish app components into directory
