+++ frontend/publish/components/Publish_stage_1.vue
