+++ frontend/publish/components/Publish_stage_2.vue
