+++ frontend/publish/components/Publish_stage_3.vue
